src/github_fetcher.py
import github
from pathlib import Path
from typing import List, Dict

class GitHubFetcher:
    def __init__(self, config: Dict):
        self.config = config
        self.api = github.Github(config['github_token'])
        
    def fetch_repositories(self) -> List[Path]:
        """Fetch relevant repositories based on config criteria."""
        # Implementation here
        pass