requirements.txt
# Core dependencies
PyGithub>=1.55
pyyaml>=6.0
pytest>=7.0
matplotlib>=3.5
seaborn>=0.11
black>=22.0
mypy>=0.950
pylint>=2.14
